delete

+++ app/Http/Controllers/TasksController.php
@@ -105,12 +105,15 @@ class TasksController extends Controller
     // deleteでmessages/（任意のid）にアクセスされた場合の「削除処理」
     public function destroy($id)
     {
-            // idの値でメッセージを検索して取得
-        $task = Task::findOrFail($id);
-        // メッセージを削除
-        $task->delete();
+         // idの値で投稿を検索して取得
+        $task = \App\Task::findOrFail($id);
 
-        // トップページへリダイレクトさせる
-        return redirect('/');
+        // 認証済みユーザ（閲覧者）がその投稿の所有者である場合は、投稿を削除
+        if (\Auth::id() === $task->user_id) {
+            $task->delete();
+        }
+
+        // 前のURLへリダイレクトさせる
+        return back();
     }
 }
